refactor: UI 프로토콜을 CommonUI 모듈로 이동
